+++ Hippo.Web/ClientApp/src/Shared/AppContext.ts
@@ -1,9 +1,10 @@
 import React from "react";
-import { AppContextShape, User } from "../types";
+import { Account, AppContextShape, User } from "../types";
 
 const AppContext = React.createContext<AppContextShape>({
   antiForgeryToken: "",
   user: { detail: {} as User },
+  account: {} as Account,
 });
 
 export default AppContext;
